Tweak to jekyll-revealjs to fix diffs
Currently:

```diff
+ a
- b
```

is rendered as:

```diff
+ <!-- .element class="fragment" --> a
- b
```

Which is a problem because the `+` here doesn't refer to a bullet point
but rather to a literal `+`, in a block where all characters have their
literal meaning (so the user actually sees that HTML comment as text).

This is a short-term fix to allow you to escape `+` as the first
character on a line.

Ironically done to facilitate a talk whose major theme is to fix this
kind of thing upstream. 😅